3th-delivery/requirements.txt
boto3
requests
psycopg2-binary
python-dotenv
tqdm
pandas
SQLAlchemy
sqlalchemy-redshift